test/deploy.js: util function for sc deployment
